# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel Users API",
        "description": "API pour la gestion des utilisateurs dans une application Laravel.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Serveur local"
        }
    ],
    "paths": {
        "/api/transactions": {
            "get": {
                "tags": [
                    "Transactions"
                ],
                "summary": "Get list of all transactions",
                "description": "Retrieve all transactions for the authenticated user",
                "operationId": "getTransactions",
                "responses": {
                    "200": {
                        "description": "List of transactions",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Transaction"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Transactions"
                ],
                "summary": "Create a new transaction",
                "description": "Create a new transaction for the authenticated user",
                "operationId": "createTransaction",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "amount",
                                    "type"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Vente produit"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Vente de produit électronique"
                                    },
                                    "date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2024-09-01"
                                    },
                                    "amount": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 150.75
                                    },
                                    "type": {
                                        "type": "string",
                                        "example": "incomings"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Transaction created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Transaction"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation errors"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/transactions/{id}": {
            "get": {
                "tags": [
                    "Transactions"
                ],
                "summary": "Get a specific transaction",
                "description": "Retrieve a transaction by its ID",
                "operationId": "getTransactionById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the transaction",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Transaction retrieved",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Transaction"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Transaction not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Transactions"
                ],
                "summary": "Update a transaction",
                "description": "Update an existing transaction",
                "operationId": "updateTransaction",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the transaction",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "amount"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Achat produit"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Achat de produit électronique"
                                    },
                                    "date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2024-09-01"
                                    },
                                    "amount": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 100.5
                                    },
                                    "type": {
                                        "type": "string",
                                        "example": "outgo"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Transaction updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Transaction"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Transaction not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Transactions"
                ],
                "summary": "Delete a transaction",
                "description": "Delete a specific transaction by its ID",
                "operationId": "deleteTransaction",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the transaction",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Transaction deleted successfully"
                    },
                    "404": {
                        "description": "Transaction not found"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Transaction": {
                "title": "Transaction",
                "description": "Modèle représentant une transaction financière",
                "properties": {
                    "id": {
                        "description": "Identifiant unique de la transaction",
                        "type": "string"
                    },
                    "user_id": {
                        "description": "Identifiant de l'utilisateur associé à la transaction",
                        "type": "integer"
                    },
                    "title": {
                        "description": "Titre de la transaction",
                        "type": "string"
                    },
                    "description": {
                        "description": "Description de la transaction",
                        "type": "string",
                        "nullable": true
                    },
                    "amount": {
                        "description": "Montant de la transaction",
                        "type": "number"
                    },
                    "type": {
                        "description": "Type de transaction, soit 'incomings' (entrées) soit 'outgo' (sorties)",
                        "type": "string",
                        "enum": [
                            "incomings",
                            "outgo"
                        ]
                    },
                    "date": {
                        "description": "Date de la transaction",
                        "type": "string",
                        "format": "date"
                    },
                    "created_at": {
                        "description": "Date de création de l'enregistrement",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Date de la dernière mise à jour de l'enregistrement",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}